updated Routes Folder and changed location of css and html files
