+++ server.js
@@ -17,12 +17,12 @@ app.use(bodyParser.urlencoded({limit: '10mb', extended: false }));
 
 
 //serve static files for css designs
-app.use('/', express.static(path.join(__dirname, '/styles')));
+app.use('/', express.static(path.join(__dirname, '/public')));
 
 
 //routes (root folder is currently the login route)
 app.use('/', require('./routes/root'));
-app.use('/home', require('./routes/home'));
+app.use('/login', require('./routes/login'));
 
 
 app.all('*', (req, res) => {
