+++ views/home.html
@@ -6,18 +6,17 @@
    <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Group 1</title>
-    <link rel="stylesheet" type="text/css" href="index.css" />
+    <link rel="stylesheet" type="text/css" href="/css/home.css" />
    
 </head>
 <body>
-
     <div class = "bookshelf"></div>
+
 
     <h1>Home</h1>
         
         <span class="psw"><a href="#">logout</a></span>
-
-    </form>
         
+    
 </body>
 </html>
